채팅 기능 호스트 추가
