+++ campingapp/src/main/AndroidManifest.xml
@@ -9,6 +9,7 @@
     <uses-permission android:name="android.permission.READ_CONTACTS" />
     <uses-permission android:name="android.permission.BIND_GET_INSTALL_REFERRER_SERVICE" />
     <uses-permission android:name="android.permission.READ_MEDIA_IMAGES" />
+    <uses-permission android:name="android.permission.BIND_GET_INSTALL_REFERRER_SERVICE" />
 
     <application
         android:name=".MyApplication"
@@ -20,6 +21,7 @@
         android:theme="@style/Theme.CampingApp"
         android:usesCleartextTraffic="true">
         <activity
+            android:name=".HostDetailActivity"
             android:name=".HostListItemActivity"
             android:exported="false" />
         <activity
@@ -113,6 +115,7 @@
             android:exported="false" />
         <activity
             android:name=".CampDoNmActivity"
+            android:exported="true" />
             android:exported="true" /> <!-- <activity -->
         <!-- android:name=".AddActivity" -->
         <!-- android:exported="true" -->
