+++ campingapp/src/main/res/layout/activity_host_detail.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ResDetailActivity">
+    tools:context=".HostDetailActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -16,7 +16,7 @@
         android:padding="10dp">
 
         <ImageView
-            android:id="@+id/back_resList"
+            android:id="@+id/HostList"
             android:layout_width="40dp"
             android:layout_height="40dp"
             android:layout_gravity="left"
@@ -51,6 +51,13 @@
                 android:layout_marginRight="13dp"
                 android:textSize="20dp" />
 
+            <TextView
+                android:id="@+id/host_name"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="10dp"
+                android:layout_marginRight="13dp"
+                android:textSize="20dp" />
 
             <TextView
                 android:id="@+id/camp_addr"
